Move backtrace/limine_module/logger/writer into utils module
