+++ xernel/kernel/src/arch/amd64/ioapic.rs
@@ -5,6 +5,7 @@ use x86_64::{PhysAddr, VirtAddr};
 
 use crate::acpi::{ACPI};
 use crate::mem::{paging::KERNEL_PAGE_MAPPER, HIGHER_HALF_OFFSET};
+use crate::{debug, dbg};
 
 pub struct IOApic {
     id: u8,
