+++ xernel/kernel/src/drivers/ps2/keyboard.rs
@@ -1,6 +1,7 @@
 use x86_64::structures::idt::InterruptStackFrame;
 
 use crate::arch::amd64::{apic::APIC, ports::inb};
+use crate::{dbg, debug};
 
 pub extern "x86-interrupt" fn keyboard(_stack_frame: InterruptStackFrame) {
     dbg!("keyboard hit");
